Github actions and tests
