+++ .github/workflows/api-canary.yml
@@ -1,0 +1,31 @@
+name: API Canary
+
+on:
+  schedule:
+    - cron: '0 12 * * *'
+  workflow_dispatch:
+  pull_request:
+    paths:
+      - 'EGG9000.Common/EggIncAPI/**'
+      - 'EGG9000.Common/Proto/**'
+
+permissions:
+  contents: read
+
+concurrency:
+  group: api-canary-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  periodicals:
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    # Advisory on PRs (a third-party outage or rate-limit must not block merge); a real failure
+    # on the daily schedule is the version-staleness alarm.
+    continue-on-error: ${{ github.event_name == 'pull_request' }}
+    steps:
+      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+      - uses: actions/setup-dotnet@67a3573c9a986a3f9c594539f4ab511d57bb3ce9 # v4.3.1
+        with:
+          dotnet-version: '10.0.x'
+      - run: dotnet test EGG9000.Common.IntegrationTests -c Release --filter "TestCategory=Network"
