+++ .github/workflows/bot-smoke.yml
@@ -1,0 +1,40 @@
+name: Bot Smoke (gated)
+
+on:
+  workflow_dispatch:
+  push:
+    branches: [master]
+  pull_request:
+
+permissions:
+  contents: read
+
+concurrency:
+  group: bot-smoke-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  smoke:
+    # Never runs for fork PRs: only same-repo refs reach the secret-bearing environment.
+    if: >-
+      github.event_name != 'pull_request' ||
+      github.event.pull_request.head.repo.full_name == github.repository
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    environment: ci-live
+    steps:
+      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+      - uses: actions/setup-dotnet@67a3573c9a986a3f9c594539f4ab511d57bb3ce9 # v4.3.1
+        with:
+          dotnet-version: '10.0.x'
+      - name: Boot bot with provided config, confirm it starts
+        env:
+          ConnectionStrings__Token: ${{ secrets.CI_DISCORD_TOKEN }}
+          ConnectionStrings__DefaultConnection: ${{ secrets.CI_DB_CONNECTION }}
+          BOT_ACTIVE: 'false'
+        run: |
+          # With BOT_ACTIVE=false the bot validates config and exits cleanly without connecting,
+          # proving the published image boots with the supplied secrets. Replace this with a fuller
+          # "connect to gateway + open DB, then exit" harness when one exists. Keep this job
+          # non-required so a live-service blip never blocks a contributor.
+          dotnet build EGG9000.Bot -c Release
