+++ EGG9000.Bot/Program.cs
@@ -72,7 +72,7 @@ try
             config.AddUserSecrets<Program>(optional: true); // Optional for Docker scenarios
         })
         .UseDefaultServiceProvider(options => options.ValidateScopes = false)
-        .ConfigureServices(ConfigureServices)
+        .ConfigureServices(EGG9000.Bot.BotHostFactory.ConfigureServices)
         .Build()
         .RunAsync();
 }
@@ -81,206 +81,4 @@ catch (Exception ex)
     logger.Error(ex, "Fatal error during startup");
     LogManager.Shutdown();
     throw;
-}
-
-void ConfigureServices(HostBuilderContext hostContext, IServiceCollection services)
-{
-    logger.Log(NLog.LogLevel.Info, "ConfigureServices Start");
-    try
-    {
-        var serviceProvider = services.BuildServiceProvider();
-        StaticLoggerFactory.Initialize(serviceProvider.GetRequiredService<ILoggerFactory>());
-
-        services.Configure<HostOptions>(options => {
-            options.ShutdownTimeout = TimeSpan.FromMinutes(5);
-        });
-
-        services.AddMemoryCache();
-
-        services.AddSingleton<DiscordQueueService>();
-        services.AddSingleton<IDiscordQueue>(provider => provider.GetRequiredService<DiscordQueueService>());
-        services.AddHostedService(provider => provider.GetRequiredService<DiscordQueueService>());
-
-        SecretsHelper.Initialize(hostContext.Configuration);
-
-        // Get connection string - supports both Docker secrets and local development
-        var connectionString = SecretsHelper.GetConfigOrSecret(
-            hostContext.Configuration,
-            "ConnectionStrings:DefaultConnection",
-            "db_connection_string");
-
-        if (string.IsNullOrEmpty(connectionString))
-        {
-            throw new InvalidOperationException(
-                "Connection string 'DefaultConnection' not found. " +
-                "Ensure it's set in: " +
-                "1. Docker secrets (db_connection_string), or " +
-                "2. User secrets (dotnet user-secrets set ConnectionStrings:DefaultConnection \"...\"), or " +
-                "3. Environment variable (ConnectionStrings__DefaultConnection)");
-        }
-
-        logger.Log(NLog.LogLevel.Info, "Using connection string from: " + 
-            (SecretsHelper.IsDockerSecretsAvailable() ? "Docker Secrets" : "Configuration/User Secrets"));
-
-        services.AddDbContextFactory<ApplicationDbContext>(options => {
-            options.UseSqlServer(connectionString, x => {
-                x.MigrationsAssembly("EGG9000.Common");
-                x.CommandTimeout(30);
-            });
-            options.EnableSensitiveDataLogging(true);
-            options.AddInterceptors(new QueryCountingInterceptor());
-        });
-
-#if !DEBUG
-        services.Configure<ActiveMonitorOptions>(options => {
-            options.ServiceType = "bot";
-            options.ColorConfigKey = "BOT_COLOR";
-            options.PollIntervalSeconds = 10;
-        });
-        services.AddSingleton<ActiveMonitorHostedService>();
-        services.AddHostedService(provider => provider.GetRequiredService<ActiveMonitorHostedService>());
-#endif
-
-        services.AddSingleton<DatabaseCache>();
-        services.AddHostedService<UserCacheRefreshService>();
-        services.AddHostedService<ActiveCoopsCacheRefreshService>();
-        services.AddSingleton<CoopStatsCache>();
-        services.AddSingleton<CoopAssignmentLookup>();
-        services.AddHostedService<CoopAssignmentLookupRefreshService>();
-        services.AddSingleton<Words>();
-        services.AddSingleton<BotLogger>();
-
-#if RELEASE
-        var release = true;
-#else
-        var release = false;
-#endif
-
-#if DEBUG
-        services.AddSingleton<Bugsnag.IClient>(new Bugsnag.Client(new Bugsnag.Configuration("0")));
-
-        var serviceCustomize = Type.GetType("EGG9000.Bot.ServiceCustomize");
-        if(serviceCustomize is not null && !release) {
-            var method = serviceCustomize.GetMethod("ConfigureServices");
-            method.Invoke(null, [hostContext, services]);
-        }
-#else
-        if(release) {
-            logger.Log(NLog.LogLevel.Info, "RUNNING IN RELEASE");
-            
-            var bugsnagKey = SecretsHelper.GetConfigOrSecret(
-                hostContext.Configuration,
-                "ConnectionStrings:BugSnagApiKey",
-                "bugsnag_api_key");
-
-            var bugsnagConfig = new Bugsnag.Configuration(bugsnagKey);
-            
-            var bs = new Bugsnag.Client(bugsnagConfig);
-
-            // Register as singleton for background services
-            services.AddSingleton<Bugsnag.IClient>(bs);
-
-            services.AddBugsnag(options => {
-                options.ApiKey = bugsnagKey;
-                options.AppVersion = Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "unknown";
-                options.ReleaseStage = "production";
-            });
-
-            var rabbitmqConn = SecretsHelper.GetConfigOrSecret(
-                hostContext.Configuration,
-                "ConnectionStrings:RabbitMQServer",
-                "rabbitmq_connection");
-            
-            services.AddOptions<RabbitMqTransportOptions>().Configure(options => {
-                var host = rabbitmqConn?.Split("|");
-                if(host?.Length > 1) {
-                    options.Host = host[0];
-                    options.User = host[1];
-                    options.Pass = host[2];
-                }
-            });
-            
-            services.AddMassTransit(x => {
-                x.AddConsumer<ShutdownConsumer>();
-                x.AddConsumer<ExpireCacheConsumer>();
-                x.AddConsumer<RestartConsumer>();
-                // Per-instance temporary queue so a version update fans out to every running process
-                // instead of being load-balanced across a shared queue.
-                x.AddConsumer<UpdateApiVersionsConsumer>().Endpoint(e => { e.InstanceId = Guid.NewGuid().ToString("N"); e.Temporary = true; });
-                if(string.IsNullOrEmpty(rabbitmqConn)) {
-                    logger.Log(NLog.LogLevel.Info, "Using RabbitMQ In Memory");
-                    x.UsingInMemory((context, cfg) => {
-                        cfg.ConfigureEndpoints(context);
-                        cfg.UseMessageRetry(r => r.Interval(3, TimeSpan.FromSeconds(5)));
-                    });
-                } else {
-                    logger.Log(NLog.LogLevel.Info, "Using RabbitMQ Server");
-                    x.UsingRabbitMq((context, cfg) => {
-                        cfg.ConfigureEndpoints(context);
-                    });
-                }
-            });
-        } else {
-            logger.Log(NLog.LogLevel.Info, "RUNNING IN DEBUG");
-            services.AddBugsnag();
-            // DiscordQueueService (and other singleton background services) require a singleton
-            // Bugsnag.IClient, which is only registered in the RELEASE branch. Register a suppressed
-            // client in Debug so local CLI runs resolve. ReleaseStage is excluded from NotifyReleaseStages
-            // so nothing is actually reported.
-            services.AddSingleton<Bugsnag.IClient>(new Bugsnag.Client(new Bugsnag.Configuration("00000000000000000000000000000000") {
-                ReleaseStage = "development",
-                NotifyReleaseStages = ["production"],
-            }));
-            services.AddSingleton<IPublishEndpoint>(new PublishEndpointMock());
-        }
-
-        services.AddSingleton<DiscordHostedService>();
-        services.AddSingleton(provider => provider.GetRequiredService<DiscordHostedService>().Gateway);
-
-        services.Configure<UpdaterOptions<LeaderboardUpdater>>(x => x.DelayStart = TimeSpan.FromMinutes(15));
-        services.AddHostedService<LeaderboardUpdater>();
-
-        services.AddHostedService<ArtifactCheaters>();
-        services.AddHostedService<StaffCoopsMessage>();
-        services.AddHostedService<CoopStatsRefreshService>();
-        services.AddHostedService<EventUpdater>();
-
-        services.Configure<UpdaterOptions<ThreadsCoopStatusUpdater>>(x => x.DelayStart = TimeSpan.FromMinutes(5));
-        services.AddSingleton<ThreadsCoopStatusUpdater>();
-        services.AddHostedService(provider => provider.GetService<ThreadsCoopStatusUpdater>());
-
-        services.AddSingleton<ContractUpdater>();
-        services.AddHostedService(provider => provider.GetService<ContractUpdater>());
-
-        services.AddHostedService<UserCXPUpdater>();
-        services.AddHostedService<NewContracts>();
-        services.AddHostedService<CreateCoopViaAPI>();
-        services.AddHostedService<CreateCoopThreads>();
-        services.AddHostedService<ShipReturnDM>();
-        services.AddHostedService<UserSnapShots>();
-        services.AddHostedService<ManageOverflow>();
-        services.AddHostedService<RemoveTempRoles>();
-        services.AddHostedService<HandleGradeChanges>();
-        services.AddHostedService<RefreshNasaApod>();
-        services.AddHostedService<UpdateBackups>();
-        services.AddHostedService<CleanAutomationLogs>();
-
-        services.AddSingleton<CoopsBeingCreatedService>();
-        services.AddSingleton<JobService>();
-        services.AddHostedService(provider => provider.GetService<JobService>());
-
-        services.AddHostedService<CommandService>();
-        services.AddHostedService<DiscordUserService>();
-        services.AddHostedService<UserGrades>();
-
-        //services.AddSingleton<IMetricServer>(_ => new MetricServer(port: botColor == "blue" ? 9464 : 9465));
-        //services.AddHostedService<PrometheusMetricServerHostedService>();
-#endif
-    } catch(Exception e) {
-        logger.Error(e, "Stopped program because of exception");
-        throw;
-    }
-    finally {
-        LogManager.Shutdown();
-    }
 }
